# api/api_server/swagger.yaml
definitions:
  api.ApiAvatar:
    properties:
      url:
        type: string
    type: object
...
  api.ApiLoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.ApiProfileUser:
    properties:
      avatar:
        $ref: '#/definitions/api.ApiAvatar'
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  api.ApiSignupUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.ApiUpdateUser:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    type: object
...
  errors.Error:
    properties:
      error:
        type: string
    type: object
...
  usecase.Session:
    properties:
      userData:
        $ref: '#/definitions/usecase.UserId'
      value:
        type: string
    type: object
...
  usecase.UserId:
    properties:
      id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Авторизация пользователя.
      parameters:
      - description: Данные пользователя.
        in: body
        name: Review
        required: true
        schema:
          $ref: '#/definitions/api.ApiLoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: Jwt auth токен успешно получен.
          schema:
            $ref: '#/definitions/usecase.Session'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Авторизация.
  /user/logout:
    delete:
      consumes:
      - application/json
      description: Выход из аккаунта пользователя.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Выход из аккаунта успешно выполнен.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Выход.
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: Получение пользовательского профиля.
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя успешно получен.
          schema:
            $ref: '#/definitions/api.ApiProfileUser'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение профиля.
    put:
      consumes:
      - application/json
      description: Обновление пользовательских данных.
      parameters:
      - description: Обновление пользовательских данных.
        in: body
        name: UpdateUser
        required: true
        schema:
          $ref: '#/definitions/api.ApiUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя успешно обновлены.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Обновление профиля.
  /user/profile/avatar:
    get:
      consumes:
      - application/json
      description: Получение пользовательского аватара.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Аватар пользователя успешно получен.
          schema:
            $ref: '#/definitions/api.ApiAvatar'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Получение аватара.
    put:
      consumes:
      - application/json
      description: Обновление пользовательского аватара.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Аватарка пользователя.
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар пользователя успешно обновлён.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Обновление аватара.
  /user/signup:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные пользователя.
        in: body
        name: Review
        required: true
        schema:
          $ref: '#/definitions/api.ApiSignupUser'
      produces:
      - application/json
      responses:
        "200":
          description: Jwt auth токен успешно получен.
          schema:
            $ref: '#/definitions/usecase.Session'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Регистрация.
swagger: "2.0"